+++ apps/nextjs/src/components/features/menus/assignee-icon-menu/assignee-icon-menu.tsx
@@ -1,5 +1,5 @@
 import { memo, useCallback, useState } from 'react';
-import { PopoverAssigneeInput } from '@/components/features/popovers';
+import { PopoverAssigneeInput } from '@/components/features/popovers/popover-assignee-input/popover-assignee-input';
 import { TeammateAvatar } from '@/components/features/teammate-avatar';
 import { Icon } from '@/components/ui/icon';
 import { Tooltip } from '@/components/ui/tooltip';
