--- apps/nextjs/src/components/features/popovers/index.ts
@@ -1,7 +1,0 @@
-export * from './popover-assignee-input';
-export * from './popover-due-date-picker';
-export * from './popover-editor-link';
-export * from './popover-emoji';
-export * from './popover-profile';
-export * from './popover-project-menu';
-export * from './popover-set-color-and-icon';
