+++ apps/nextjs/src/components/features/popovers/popover-project-menu/menu-list.tsx
@@ -1,4 +1,4 @@
-import { PopoverSetColorAndIcon } from '@/components/features/popovers';
+import { PopoverSetColorAndIcon } from '@/components/features/popovers/popover-set-color-and-icon/popover-set-color-and-icon';
 import { ColorBox } from '@/components/ui/color-box';
 import { Flex } from '@/components/ui/flex';
 import { Icon } from '@/components/ui/icon';
