+++ apps/nextjs/src/components/features/projects/project-tile-item/project-tile-item.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react';
-import { PopoverProjectMenu } from '@/components/features/popovers';
+import { PopoverProjectMenu } from '@/components/features/popovers/popover-project-menu/popover-project-menu';
 import { TeammateAvatar } from '@/components/features/teammate-avatar';
 import { AvatarGroup } from '@/components/ui/avatar';
 import { Fade } from '@/components/ui/fade';
