+++ apps/nextjs/src/components/features/task-detail/task-detail-body/form/subtasks/tasks-name/due-date/due-date.tsx
@@ -1,5 +1,5 @@
 import { memo, useCallback } from 'react';
-import { PopoverDueDatePicker } from '@/components/features/popovers';
+import { PopoverDueDatePicker } from '@/components/features/popovers/popover-due-date-picker/popover-due-date-picker';
 import { DueDate as AtomsDueDate } from '@/components/ui/due-date';
 import { Icon } from '@/components/ui/icon';
 import { Tooltip } from '@/components/ui/tooltip';
