+++ apps/nextjs/src/components/features/tasks/tasks-list/tasks-list-cells/tasks-due-date/tasks-due-date.tsx
@@ -1,6 +1,6 @@
 import type React from 'react';
 import { memo, useCallback, useMemo } from 'react';
-import { PopoverDueDatePicker } from '@/components/features/popovers';
+import { PopoverDueDatePicker } from '@/components/features/popovers/popover-due-date-picker/popover-due-date-picker';
 import { TasksListCell } from '@/components/features/tasks/tasks-list/tasks-list-cell';
 import { DueDate } from '@/components/ui/due-date';
 import { Flex, type FlexProps } from '@/components/ui/flex';
