+++ apps/nextjs/src/components/pages/home/components/tasks-due-soon/list-item.tsx
@@ -1,7 +1,7 @@
 import type React from 'react';
 import { memo, useCallback } from 'react';
 import { ProjectChip } from '@/components/features/chips/project-chip';
-import { PopoverDueDatePicker } from '@/components/features/popovers';
+import { PopoverDueDatePicker } from '@/components/features/popovers/popover-due-date-picker/popover-due-date-picker';
 import { CheckIcon } from '@/components/ui/check-icon';
 import { DueDate } from '@/components/ui/due-date';
 import { Flex } from '@/components/ui/flex';
