+++ apps/nextjs/src/components/ui/editor/editors/node-views/mention/task/task-link.tsx
@@ -1,10 +1,8 @@
 import { memo } from 'react';
-import {
-  PopoverEditorLink,
-  PopoverEditorLinkContent,
-  PopoverEditorLinkText,
-  PopoverEditorLinkTrigger,
-} from '@/components/features/popovers';
+import { PopoverEditorLink } from '@/components/features/popovers/popover-editor-link/popover-editor-link';
+import { PopoverEditorLinkContent } from '@/components/features/popovers/popover-editor-link/popover-editor-link-content';
+import { PopoverEditorLinkText } from '@/components/features/popovers/popover-editor-link/popover-editor-link-text';
+import { PopoverEditorLinkTrigger } from '@/components/features/popovers/popover-editor-link/popover-editor-link-trigger';
 import { CheckIcon } from '@/components/ui/check-icon';
 import { useProjectTask } from '@/store/entities/project-task';
 import { useTask } from '@/store/entities/task';
